Modified strucutre of getting enteties
